Refactored RPC::Connection to be a little bit transport agnostic
This allows it to talk to things like bitcoind. It still wants a tcp
socket as its transport -- but it can at least now be wrapped in HTTP
for bitcoin as well as re-use of the same code for ElectronX style
linefeed based JSON.

Our RPC lib is nowhere near 100% generic and re-usable but it's a good
first step here with this refactor job.