Added ParseHexFast() to the Util namespace
This hex parser is 5-6x faster than what we were using before, although
it does no real verification that the input is hex digits.  Usable with
results returned from bitcoind, which is never going to be
malicious/malformed and thus we can rely on this non-checking hex
parser for the performance it offers (at the expense of verification).