Redid the hashXAggregated structure in BlockProc
It's now an unordered_node_map, keyed off of HashX.  This should be more
useful down the processing pipeline.  What we had before was confusing and
difficult to use.. its only advantage was it preserved order of HashX's
in a block.

We can reproduce the order though on-demand if it turns out we need it
by adding a method to construct an array keyed off the in/out index per
hashX.

So this map approach is sufficient and likely more useful.