Regined the backlog avoidance mechanism slightly
Tightened the lock and made it a shared_lock to hopefully stall the dl
tasks less waitng for it.

Also moved all logic for it into the method that decides if throttling
is necessary.

This mechanism is still very awkward. TODO: figure out something more
elegant.  It's just a PoC for now.