Tweaks and fixups in storage class -- refactored some stuff
We no longer store the headers in memory. We keep them in the db.  We
don't need them that often and clients can just read them from the db
if/when needed. Hopefully rocksdb is peppy enough.  We can cache them if
it's a problem.

Also refactored some stuff around and simplified the storage code a
little bit.