Reduced CPU usage in addBlock -- we mark outputs spent in PreProcessedBlock now
This reduces the building of a set for each hashX in addBlock. Now we
flag the outputs that are safe to ignore for adding to utxo set (because
they were already spent) by flagging the OutPt directly in
PreProcessedBlock fill().

Should improve performance slightly as no more nested loops in
addBlock()