Added saving of undo data to db
The save is done for the last 10 blocks before latest tip.

We also have to unconditionally issue Delete commands to the db for all
block heights >10 blocks ago.  So far the unconditional issuance of the
delete command does not seem to impact performance.

We could mitigate the Delete command by keeping track of the olders
header we know exists in the db -- but we don't bother for now.