DB changes, phase 1. Added 'dirty' flag, added utxoct to meta table, ...
We skip the utxo count check at startup unless the new -C (--checkdb)
flag is set, in which case we do a very thorough check of all utxos and
the corresponding scripthash_unspent entries.

We set a dirty flag during block addition / removal and if it's set, we
abort the app immediately and ask the user to do a full resynch.

Since adding/removing blocks isn't 100% ACID, this is needed until we
figure out how to redo the data to make everything ACID.

Also in this commit -- attempted to see if I could make the header
reading faster from the db by using MultiGet.  It turns out it doesn't
help much.