Reduced the frequency of mempool status printing to the log
The rule is: debug mode prints every time it changes to debug logger.
Normal logger prints at most once every 30 seconds, and only if the
mempool is non-empty and if the size has changed, and if "up-to-date"
(in Controller).

The logic is a bit convoluted, but at least it doesn't print too often.