Bugfix for Linux: It turns out timer events always arrive after signals
This is on Linux only.  Who knew?  Thanks for Jt for finding this bug.

We had to go ahead and get rid of the Util::AsyncOnObject() call in
putBlock and lieu of a signal/slot connection (putBlock connected to
protected on_putBlock).

This fixes the race condition observed on Linux only whereby blocks
would be done downloading but the download task would get deleted before
the blocks could get processed by the Controller.

By making all of the events the Controller listens for arrive via
signals now, we avoid any out-of-order issues since signals themselves
always get delivered in order (whereas it appears timers get processed
AFTER the signals but only on Linux! On windows and macos they do not!).