Tweaks to Server.cpp -- grab daemon version only once
We were grabbing it with each RPC call to server.banner, which is pretty
inefficient considering it involves a back-and-forth to the daemon and
the daemon doesn't even change that often.

Instead, we detect when it changes and cache the 2 pieces of information
we need from it (version number and subversion string).

We attach to BitcoinDMgr's `gotFirstGoodConnection` to detect a
reconnect and then send it a request for "getnetworkinfo".