Fix for Linux w/ PeerMgr: AsyncOnObject strikes again! Signal race condition fix
On Linux it appears soemtimes the PeerMgr "allServerStarted" would end
up executing in the wrong thread because the PeerMgr didn't end up
living in its own thread in time.

Instead we do it the right way by waiting for its thread to actually
start executing as a result of startup() and we use a signal/slot
connection from SrvMgr -> PeerMgr to tell it when all the listening
servers are up.

This should fix the warnings seen on startup by Axel Gembe.