Build system rejection of Qt 5.12.4 or earlier and Qt 5.13.0
See this bug: https://bugreports.qt.io/browse/QTBUG-76276?

This is a fairly serious bug in Qt that could cause Fulcrum to crash
randomly and/or could lead to db corruption.  The bug has been fixed in
Qt 5.12.5 and Qt 5.13.1 (or above).

We must disallow building with affected prior versions as that would
lead to a poor experience for admins.

The bug involves the DNS resolver calling your code back in the wrong
thread (see link).