Refactored out the Util::ThreadPool globals into a class
This class is currently a singleton, but it no longer uses the internal
Qt QThreadPool::globalInstance(), but rather a custom QThreadPool
instance we create for our app.  The reason being that Qt uses the
global QThreadPool instance itself for DNS lookup and other tasks and we
do not want our thread pool to interfere with Qt (or be interfered with by Qt).

This ensures we have finer grained control over the ThreadPool dedicated
to processing Fulcrum requests.