Added a comment in Storage.cpp to remind myself to optimize getHistory
There is low-hanging optimization fruit in that method that I need to
spend some time picking.  For now, the existing getHistory is "fast
enough" in that it's faster than the ElectrumX/ElectronX version of the
function -- but its performance could be significantly improved by
writing batched versions of the txhash/txpos lookup functions
hashForTxNum and heightForTxNum.