Bumped the pending conneciton limit to 60, and a nit in Servers.cpp
Default pending connection limit for QTcpServer is 30 (which is already
pretty generous for a sufficiently asynchronous server).  We bumped it
to 60 just to be paranoid.

Also after careful reading of the QTcpServer docs, I realized that if
the limit is reached, the newConnections signal is not emitted but
many connections can still be pending beyond the limit.

We rewrote the slot for this signal to process as many pending
connections as are available in a loop.

Minor nit, but it may make a difference someday on a loaded server.