Made the AdminServer use its own custom small threadpool
Some of the admin RPC commands will need to block as they call lambdas
on objects and wait for results.  We will do this in a small threadpool
of 2 threads that is our own custom pool for just the admin rpc
instances.  This is so that admin RPC calls, even if expensive, will  not
interfere with the rest of the app and very importantly will not
interfere with regular SPV client requests.

Also added some stats info for the admin rpc server to the /stats data.

Also in this commit: some refactoring.