Admin RPC: Added the 'addpeer' command
Basically enqueues the peer immediately if not already in the peer list.
If this is an "update" to an existing failed peer, it will update the
port info and the app will try again in ~10 mins.

If the peer already exists and is already valid and connected, then the
addpeer request is silently ignored (even if the ports specified on CLI
differ).

Coming soon: 'rmpeer' to explicitly delete peers (but not ban them)