Client class: Ignore new messages when disconnecting
This will save some cycles for clients that are being disconnected
(either due to protocol error, bad version, or dropped connection) -- we
short-circuit ignore new messages arriving after the decision to drop a
client has been made.

This should save cycles and reduce processing for a client that is going
away anyway soon.