Bugfix: listunspent now correctly deletes confirmed spends in mempool
Oops! This is entirely to my not reading the protocol specification
carefully and because Electron Cash doesn't even use this method, so
it slipped through the cracks in testing.

It turns out listunspent should delete confirmed spends from the
returned results (that is, it should have the mempool tx's SPEND
confirmed utxos and not include them in the result set).  It wasn't doing
that properly.  This commit fixes that.