Added auto-zombie reaping when global subs limit reached
This is activated 125ms after kicking the most-subscribed client to
reduce the size of the subs table immediately since we are at the limit.

This code hopefully allows for subsequent clients to continue to use the
server by cleaning up the resources of the kicked client ASAP.

Also in this commit -- some nits/fixes.