SubsMgr: Small performance optimization for cases where network is busy
Added a branch/conditional to the SubsMgr `doNotifyAllPending` to
optimize the way we loop.  We now always loop over the smaller of the two
sets: pendingNotifications vs. subs.  This may save cycles in
circumstances where tens of thousands of addresses have changed at once
(this normally only happens if blocks are nearly full).

Our code now copes well with both low-BCH-network-usage and
high-BCH-network-usage circumstances and takes the faster path in either
case.  (Previously it was optimized for low-BCH-network-usage only).