Small optimization: Use QStringLiteral in often-called places
QStringLiteral avoids a QString() constructor (it instead builds the
QStringData at compile-time).  In some key places in the code we
replaced QString("somestring") with QStringLiteral("somestring").  This
is a tiny optimization to shave CPU cycles off, and slightly sacrifices
readability so we only do it when it matters.

Also in this commit prefer split('c') versus split("c") for similar
reasons (only really matters in hotspots).

We also made the LineFeedConnection use a static QByteArray for the
"\r\n" string as opposed to parsing the "\r\n" C string each time.

These are tiny performance nits, really, but they enforce good practice
going forward.