Improved connection pipeline in ServerBase
We now check the Per-IP limit early on in the connection pipeline in
ServerBase, and if the would-be client exceeds the connection limit, we
abort the connection ASAP, rather than proceeding.  Previously we would
create Client * instances for such connections, only to delete them
right away in SrvMgr where the Per-IP limit detection was taking place.

Moving this detection limit earlier in the pipeline is not only more
efficient and more correct, but it is also a requirement for the upcoming
WebSocket implementation, which has a potentially longer handshake process
and which best benefits from rejecting spammy clients as early as
possible.

Also, in this commit we don't `addPendingConnection` in ServerSSL until
*after* a successful SSL handshake has occurred.  This prevents the
creation of Client * instances until later, after we actually have a
fully established encrypted connection.  This is more correct since we
can't even communicate with the Client until the handshake has taken
place, and the way it worked before only worked by accident since we
don't send any data to clients until they send to us first.

Note that this change as well will be needed for the upcoming WebSocket
implementation.