Refactored the incomingConnection code in ServerBase and ServerSSL
We had two parallel implementations for both Server and ServerSSL's
incomingConnection override. This has been refactored into basically some
common code that these overrides call into.  The initial limit checking,
the WebSocket portion of the handshake (if applicable), as well as misc. error
checking, are now all implemented in ServerBase.

ServerSSL calls into this common implementation after the TLS part of the handshake
is complete.

Regular (TCP) Server also uses the same code for the WebSocket portion of the
handshake as ServerSSL (if it's a web socket server).