Fixup - made RPC::Message::Id a real type and not a QVariant
RPC::Message::Id was an alias for QVariant, but this was not ideal as it
was 1. a heavy type and 2. wasn't clear if it was safe to use in a QMap
as a key.

So.. we implemented a real type for RPC::Message::Id, which is a very
tightly constrained variant that can either store Null, a QString, or an
int64 and nothing else.

We also changed the idMessageMap to a QHash now.