Floating point fixup for RPCMsgId
We squash id 1.0000000 down to "1", and this check was taking a
difference of the fractional part to ensure it's zero.  Strictly
speaking the check should not be comparing <= 0.0, but should be
taking a difference and ensuring it's <= Epsilon (smallest fraction from 0 -> 1).

This is because doubles may become subnormal.

The code as been fixed to handle the situation correctly.