Template fix in Util::keySet and Util::valueSet
This is a nit-commit. The default for keySet and valueSet, if no arg is
specified, is to return a std::set. However, we were calling .reserve()
on this which would produce a compile-time error if caller used the
default keySet() or valueSet() template.  This has now been fixed to
correctly use SFINAE to only call set.reserve() if that method exists
(e.g. for std::unordered_set).