Optimization for BitcoinD client - reduce mallocs and overhead
We pre-allocate the entire payload buffer that we will send to bitcoind
in wrapForSend(), including headers. This reduces redundant mallocs and
saves CPU cycles.  Previously we were using a QTextStream writing to a
QByteArray and this was not ideal not only for the potential for many
mallocs, but also because QTextStream does a round-trip to QString and
back via toUtf8() when appending QByteArrays.  This was unnecessary
since we are guaranteed to be encoding utf8 strings in our header since
it mostly comes from static or otherwise ascii strings.

Anyway, this commit optimizes HttpConnection::wrapForSend() slightly to
do 1 big malloc at the beginning and then just write to that buffer,
saving CPU cycles for our BitcoinD client, as well as not use
the slightly slower QTextStream.