Fixed .toDouble() and .toLongLong() and .toULongLong() in parser
- QByteArray's .to*() functions may take a copy of the string (which is
just an extra heap allocation).  We want to avoid this so we use
std::array on the stack and make a temp copy of the C string we are
parsing to turn into a number, then we call the C function strtoll, et
al in order to parse the JSON numeric. This is faster than calling Qt
and leads to less heap fragmentation.

- Added more corner cases to the JSON tests.