[wip] Refactor away some often-user boilerplate for bytes into ByteView
When implementing the RollingBloomFilter (and the TrivialHashHasher)
there was ugly boilerplace, reinterpret_casts, and support for various
POD types as well as QByteArray, std::vector, etc.  All of this has been
refactored into a utility class, ByteView, which has templated
auto-conversions from often used containers, e.g.: base_blob, QString,
QByteArray, std::vector, etc.

This class is heavily constexpr and noexcept (so hopefully it compiles
away into nothing).  It is lightweight and inherits from
std::basic_string_view but its underlying type is std::byte * (rather
than char *).

It is intended to encapsulate a pointer, size pair into binary data
blobs.

This is a work-in-progress since I'm not sure all this heavy template
magic is worth it in the long-run.