More bchd quirk fixups and refactoring
- Moved all logic for querying bitcoind version into BitcoinDMgr where
it belongs.  We keep a single copy of the BitcoinDInfo object now, and
the servers grab a copy of this object in a thread-safe manner when they
need information from it.

- Added a specific code path for parsing the "version" integer for bchd.
Bchd's format for this number is somewhat strange and we do our best
here. TODO: Write to Josh and Chris Pacia and figure out how to really
parse this number properly.

- Added bitcoind version info to the /stats enpoint and 'getinfo' admin
rpc calls.