Use BitcoinD's genesis hash to determine sanity, plus bugfixes
- Whenever we (re)connect to bitcoind, we now refresh the hash of block 0.
We compare this hash to the hash stored in the DB and if the two
disagree, we quit the app (likely the user is pointing this instance to
the wrong chain).  Previously we used the "chain name" reported by
bitcoind for this sanity check -- but seeing as how that string is not
standardized -- it's safer to check against the genesis hash.

- A bug fix:
  - Division by zero could happen in the admin rpc 'getinfo' when
calculating avg. fee rate. The resulting number was basically infinite
nonsense. This has been fixed.

- Nits:
  - use {} notation in a few places
  - replace constexpr with inline constexpr for global constants
  - reduce emplace_back abuse
  - changed literal doubles to use trailing 0 e.g. 100.0 as opposed to 100.
  - Improved error messages on getnetworkinfo and getblockhash failures
in BitcoinDMgr
  - Print "refreshed" messages to debug log when we grab a new
subversion and genesis hash from bitcoind on (re)connect
  - Mempool.cpp was missing the license preamble.