ByteView template class fixups
- Got rid of all constexpr methods -- they relied on casting through
void * which is not allowed in constexpr. Even though the compiler
didn't complain, casting through void * in a constexpr is not portable
across all compilers, since the standard says it's forbidden. In practice
the lack of constexpr doesn't cost us much -- the optimizer will always
optimize ByteView away to being a zero-cost abstraction if using -O2 or
above.

- Made ByteView work implicitly with std::array of POD data.

- Made ByteView template c'tors reject C arrays that are arrays of pointers.