Broadcast rpc tweak - save txkey to both lambdas
We are forced to calculate the txkey as a cheap hash for the tx bytes
anyway -- so we might as well save it to both lambdas and use that as
the key for both the broadcast success and broadcast fail bloom filters.

Also removed some dead code (unused txkey.isEmpty check).