BitcoinDMgr: simplified logic for getBitcoinD() (should be faster)
The logic for getBitcoinD() was unnecessarily slow in that it would
generate a new random number for a random bitcoind in the goodSet each
time.  Instead we are just going to use a simple round-robin strategy to
dole out bitcoind pointers to submitRequest(). In practice this is much
faster and shaves cycles off the creation of new bitcoind requests.

Also in this commit: defensively clean up the reqContextTable in
on_finished() (which runs right before the BitcoinDMgr thread is about
to end).