bchd support: added --bitcoind-tls option
This option can be specified on the CLI as `--bitcoind-tls` or in the conf
file as `bitcoind-tls = true`.  If enabled, then we will connect to the
remote bitcoind using TLS (https) rather than a bare TCP socket (http).
This allows Fulcrum to work flawlessly with default bchd setups, which
enable TLS for all HTTP connections by default (unless the `notls` option
is given to bchd).

See issues #43 and #28.