bchd: Got rid of "quirks" workaround code in BitcoinDMgr
Since the first release version of bchd, 0.16.5, is one we can talk to
*without* any quirks (thanks Josh!), we can eliminate the code path in
the bitcoindmgr request code that tries to rewrite the request to
bitcoind to deal with bchd-specific quirks.  This should save CPU
cycles when synching bitcoind.

The "zeroArgEstimateFee" quirk workaround for BCHN and ABC was moved back
into Servers.cpp just for that RPC method, so it doesn't affect the
common-case call path for *all* bitcoind RPC requests.

This all shaves cycles off the bitcoind requests and reduces CPU overhead
slightly when synching.

Also in this commit: use QStringLiteral("ping") rather than bare C
string "ping" (saves cycles).