Added --wss-cert and --wss-key CLI and corresponding conf args
These are so that WSS ports may use CA signed certificates, whereas
legacy Electrum Cash ports can continue to use the old pinned
self-signed certificates.

This is needed because older Electron Cash clients refuse to connect to
a server if it is using CA-signed certs if it had previously seen a
self-signed cert from that server. The bug is fixed in Electron Cash
master, but we want to allow admins to set up WSS using CA-signed certs
(because this is what browsers require), while still preserving the old
self-signed certs for legacy EC -t/-s ports.