Fixed issue where TXO type was not being hashed properly
This closes issue #47.  We introduced a regression in one of our
previous versions where we switched over to using the ByteView type for
hashing.  The TXO type was stuffing its data into a temporary struct and
that struct was being hashed, but that temporary struct did not have a
unique object representation, leading to bugs.

This affected the `listunspent` RPC call.  This commit fixes that issue
and the `listunspent` call works again.

Closes issue #47.