Docker: Preload jemalloc to reduce heap fragmentation
Jemalloc is better suited for the memory usage pattern of Fulcrum,
which is allocating/freeing lots of random sized buffers from multiple
threads, resulting in lots of fragmentation.

Jemalloc is a general purpose malloc(3) implementation that emphasizes
fragmentation avoidance and scalable concurrency support.

gdb was used to check the malloc symbol comes from libjemalloc.so.1