Add jemalloc static library to the static library builder for Windows & Linux (#56)
This makes the docker-based static library builder (`contrib/build/build.sh`) 
clone and compile jemalloc. It then links the jemalloc library statically the 
generated static executable.

Tested and works on both Linux and Windows.

Also in this commit: Added a "Jemalloc" key to the stats outputted by the 
"/stats" HTTP endpoint, which includes some of the output of `malloc_print_stats()` 
from Jemalloc.  Note that for this to contain any data, the jemalloc library must be 
linked at compile-time.