Fixups and tweaks to the stand-alone rocksdb_perf_memusage_test
This test is a stand-alone program that tests performance for writing
and reading ot a dummy rocksdb database.  It also is a good way to see
just how much memory rocksdb uses during normal operation.

I modified this test a little bit to use multiple threads when reading
and also I cleaned up the code a little bit.