Nit in Storage.cpp to more safely handle the raw rocksdb::DB pointer
In the error case, it would have been theoretically possible to leak the
generated pointer if the rocksdb API returns error status but does
allocate the DB object.  We more defensively handle this condition now
by immediately placing the DB * into a temporary unique_ptr before
moving it to the final unique_ptr in which it will live if there was no
error.