Allow Fulcrum to synch to bitcoind even if bitcoind is in IBD
I initially wrote fulcrum to not allow synching while bitcoind was still
in "initial block download" in order to keep the Controller code simple,
since I was still busy debugging the Controller task, and I wanted as
little complexity there as possible.

Now that the Controller task code is mature and known to be bug-free, it
was not hard to enable synching while bitcoind is still in IBD. The code
changes were not significant to manage the situation appropriately.

This change allows new installs to get up-and-running quickly since it
reduces the time that an admin has to wait before his server is ready
from: `IBD_time + fulcrum_synch_time` to the guaranteed shorter
`max(IBD_time, fulcrum_synch_time)`.