Redid BTC vs BCH auto-detect logic, got rid of --btc CLI arg
The --btc CLI arg is superfluous.  We can just detect the coin based on
user agent.  Unknown user agents for bitcoind default to BCH.

This implies that we cannot use BTC with anything other than Satoshi as
the bitcoind, which is fine for now.

TODO: Possibly support more BTC bitcoinds in the future, if it proves
that Fulcrum becomes popular on BTC.