RocksDB: Check if C++ interface is exported by the system library
Shared Windows builds of RocksDB produce a .dll file that only
dllexports the C interface of RocksDB, which will cause the config
tests to use the system library but link will fail because the C++
symbols are undefined.

This patch makes the config test link to the C++ interface to check
if the system library exports the C++ symbols.