Fix mempool synch performance when tx's are dropped (#67)
This massively refactors the `SynchMempoolTask` and the `Mempool` class to 
support dropping txs in a more performant manner.  Gone are the days of 
necessitating a full mempool re-download when a single tx is dropped.  Instead,
we now undo the effects of a mempool tx when dropping.  Note that dropping a tx
implicitly drops all of its descendant txs as well.  In practice this is an ok
limitation.

This change is particularly important for supporting BTC, which frequntly has 
single tx's replaced by RBF double-spends.

Also congested mempools (such as was seen recently on ABC chain) should do
much better now since we don't have to rebuild the whole mempool each time
for a single dropped tx.

Also in this commit: 

- Bumped version to 1.3.2
- Fixed a bug where the TXO type had a broken `operator<` (this operator was
thankfully never called)
- Added a "mempool" bench (which is really more of a test).  This test heavily
tests the new `dtopTxs` call to ensure it is sane.
- Various small bits and bits of refactoring.

---

Squashed commit of the following:

* Initial commit, work in progress.

* Added more tests for the new Mempool::dropTxs, fixed a bug in struct TXO

- Tests now are in the "bench" for mempool (since they require external
  data that I do not wish to commit to this repo).
- struct TXO had a bug -- operator< was wrong. Fixed.  Luckily this was
  never called anywhere in the codebase outside of the tests we just
  added.
- Added support for modifying the currency unit used by Amount::ToString()
  and CFeeRate::ToString()

* Moved tne mempool json dump code to Mempool.cpp

Also added an extra paranoia check when doing addTxs in Mempool.cpp to
catch bugs if we ever change the code.

Added debug dump of mempool as we bench/test towards the end so I may
visually inspect what is happening.

* Fixups and integration with the Controller class

Controller now drops txs using Mempool::dropTxs() rather than clearing
mempool each time.  Still a work in progress and still testing but so
far correctness looks like it has been tentativel achieved (more testing
is needed though to be certain).

* Compile fix for G++ 7.5.0

Meh.

* Some nits

* Rehash mampool hash tables on new block

Just to be a bit tidy with memory.

Also in this commit: remove Mempool::clear() implementation from the header
and put it in the .cpp file.

* Fixed a typo

* Updated a debug message to refer to dropTxs not addBlock

* Misc refactor and also added proper testing for "drop any tx" mode

* Tweaks and minor refactor to SynchMempoolTask

* Bumped version to 1.3.2

Also default build without ENABLE_TESTS again.

* SynchMempoolTask: Release lock earlier

Also some comments updated, dead code removed, and reserve tx hash table
size to the anticipated size to save on reallocs.

* Updated a comment

* Clear mempool on block undo.

The performance hit there is the lesser of two evils.