Mempool: Added extra paranoia check that txvec invariant is not violated
The `txvec` for a particular hashX in the mempool hashXTxs map should
always be non-empty (otherwise the hashX has no business being mapped!).
Check this is the case in Mempool::rmTxsInHashXTxs, and warn if the
invariant is violated (should never happen).

Also added this check to the mempool bench test.