Allow for >65535 TXO IONums
Previous to this commit, we serialized TXO outN (IONum) as a 16-bit
number to the database, which theoretically is not enough. It's possible
for a bitcoin tx that is 1MB to have ~111k outputs (if they are all empty
scripts).  When I first wrote Fulcrum and designed the database format,
I did not realize such a thing is possible.

Luckily, no such tx's exist on any of the 6 chains we support between
BTC and BCH, however we should be able to handle such tx's should they
appear in the future.

This commit expands on the db format in a backward-compatible way, by
extending the TXO and CompactTXO formats to optionally serialize a
3-byte IONum if the value is >65535 (2-byte if <= 65535).

Internally now IONums are 32-bit numbers, but they only get serialized
as 16-bit numbers in the common case and 24-bit numbers in the rare case
of outN being >65535 (which is enough given current bitcoin consensus rules).

We were also forced to change the Undo format from v1 to v2 as a result
of this change. Older Fulcrum will not be able to use the undo table
written by newer Fulcrum.

This commit does introduce "breaking" changes to the DB format -- DBs
written by Fulcrum after this commit are not guaranteed to be 100%
readable by older Fulcrum.  However the new format is guaranteed to be
correct whereas the old format was not.

In practice it is hoped that server admins won't notice this "breaking"
change since I don't believe admins ever downgrade their Fulcrum
versions -- they tend to always move forward in terms of version.

And if a user were to read on old db with this newer Fulcrum it would be
fine.  The potential break only can happen if going backwards in terms
of Fulrum version after having run a version including this commit. (And
even then it likely won't get noticed unless there is a reorg right
after the db is loaded, in which case the undo info would not be readable
by older Fulcrum).