Docker static build: Use Qt 5.15.2
We updated the Dockerfile to use Qt 5.15.2 for the Windows and Linux
static build.  The old ubuntu16 build however must remain on Qt 5.14.2
for now.  5.15.2 won't work there because it requires openssl 1.1.1
which is lacking in Ubuntu 16.  We could build openssl ourselves, but
that can be done at a later time.  For now the older linux build will
just use an older Qt, it's fine.