Added VarInt and Span utility classes
VarInt - Serialize any integral type to a compact format. Useful for
saving disk space when most of the time the integer you are storing
is small, but needs a large range. Future planned work may use this
class for the db.

Span - A utility template class that implements a subset of C++20's
std::span. This class is mainly taken from BCHN sources but adapted a
little bit to suit my tastes.

Also added some tests for the above.

Neither of these classes are yet used by the codebase but they may be
in the future.