Disabled dropping of dsps if they disappear from the node's dsp list
We only drop dsp's when the associated owning tx for them goes away (is
confirmed or evicted).  In this way, dsps follow the lifecycle of their
owning tx.

The reason we don't drop them if the node reports them done is that the
node can be flaky and can get restarted by the admin.  We want to store
dsps we knew about until the associated tx is gone from mempool.

This is better to avoid corner cases.