Optimized status hash calculation code + added statushash test
We were using a QTextStream before to write out the status hash to a
QString, only to convert this QString immediately to a utf8 QByteArray.
For very large histories, this would be quite wasteful. It would involve
excess allocations and copying.

Instead, we optimized the status hash calculation code to pre-allocate a
desination buffer ahead of time (sized to the anticipated worst-case size
of the history string).

We then use a stack-based formatting buffer to format the history items
individually. This code was hand-tuned to be fast.

In measurements the new way of calculating a status hash is 2x faster
than the old way for large histories.  Likely it also parallelizes
better due to fewer allocations (and thus less lock contention into the
allocator).

This commit also fixes the type "Util::AsyncSignalSafe::SBuf" to be more
correct / handle corner cases better.

This commit also adds a utility function Util::getRandomBytes() to just
generate random bytes (using Qt's QRandomGenerator).

Also added a test: "statushash" which compares the old way vs the new
way for both correctness and speed.

The old way also seemed to crash on histories larger than 10M (bugs in
Qt?).  The new way has no such problem.