Optimized "SubsMgr::getFullStatus()" further
Previous to this commit, getFullStatus would pre-allocate the entire status
string buffer ahead of time when calculating the status hash. This not only
involved extra allocations that were unnecessary (thus wasting cycles,
fragmenting heap, reducing parallelism, etc), but it was also prone to
potential OOM/std::bad_alloc due to the fact that on large histories the
status string could be GIGABYTES.

Instead, we just construct a bitcoin CHash256 hasher and we update the hasher
in-place for each element in the history and at the end finalize it.  This
should serve to reduce load on servers as well as allowing for arbitrarily
large histories.

Previous to this commit we were limited in how large a history we could
hash (~27M items). After this commit there is no theoretical limit.