Simplified the dsp tx notification logic
We just build a big set of txids and notify on that. Before what we had
was needlessly complicated, bugprone, and wasn't even correct.

The new way is simpler and correct -- just notify on all txids involved
in all touched dsps.

This is efficient and captures all state changes.

Note that now we don't immediately unsubscribe dsp subs for txs leaving from
mempool.  Instead, we let them expire and get picked up by the expire
expireSubsNotInMempool timer in DSProofSubsMgr.  This allows for keeping
subs "sticky" in the event of reorgs, and also further simplifies all
the logic.