Added "allInputsSpendP2PKH" flag for mempool txs
This is tracked by examining the scriptSigs of each tx's inputs and then
seeing if each scriptSig matches the canonical scriptSig for a p2pkh spend
(script heuristic: <push sig><push pubkey>).

We extract the pubKey from the scriptSigs and then compare the pubKey to the
scriptHash for the cached utxo (we must do this because we only save
scriptHashes and not actual scriptPubKeys in our utxo set).

If the scriptHash compares equal, we conclude the input is a p2pkh spend.

If all inputs to a tx are p2pkh spends, then we set the
"allInputsSpendP2PKH" for the mempool tx to true.

This flag will later be used to track DSP scores for mempool txs.