Added --compact-dbs CLI arg
This option compacts all of the databases on startup, which reduces disk
space consumed by the DBs.  Note that Fulcrum already compacts databases
in the background periodically, so this option is not strictly
necessary. However, it may occasionally prove fruitful to use this
option in some circumstances, or for testing.