Fixed a spurious "FIXME" warning that actually is not a bug.
This warning can occur if a DSP has descendant txs with a diamond
pattern. This is very rare, but if such a DSP occurs, the log would
print a warning message that it failed to add a tx<->DSP association
with a "FIXME" attached to it.

This situation is not a bug, but rather a rare occurrence. So we demoted
the message to "Debug" from "Warning".

It is only a bug if this is the first input in the tx in question, in
which case something is very wrong (and we do warn in that case).