Fixed a spurious "INTERNAL ERROR" warning that actually is not a bug.
On occasion, a peer may disconnect from us before we can attach a per-IP
object to it in attachPerIPDataAndCheckLimits. In that case, the
socket->peerAddress() will be null, which leads to a spurious
"INTERNAL ERROR" message printed to the log.

Since this is not actually an internal error, the message has been
downgraded from a hard error to a warning, minus the "INTERNAL ERROR"
prefix.