Build: Reject system rocksdb if not compiled with RTTI
This makes `config.tests/rocksdb` check if the library was compiled with
USE_RTTI=1 by creating an instance of an `AssociativeMergeOperator`
derived class. When rocksdb was compiled without RTTI this should
produce

```
undefined reference to `typeinfo for rocksdb::AssociativeMergeOperator'
```

when running `qmake` in `config.log` and use the static library instead.