Added `bchn.getdsproocscore` rpc method (not scalable!)
This is for private servers only. This is discouraged to be used on a
public-facing server since all this does is forward the call to BCHN,
which is a DoS vector since BCHN doesn't cache the information and the
call is potentailly slow if the mempool is set up in a pathological way.