Made _send* functions in RPC::ConnectionBase no longer virtual
Nothing in the codebase was overriding them, so they are best left
non-virtual. In the original design of the code I intended for
subclasses to customize these functions, but the design of the system
went in a direction where lots of behavior is controlled by
RPC::ConnectionBase anyway.

Making them non-virtual should hopefully shave a few cycles when
processing messages.