Auto-transform author's address for BTC/BCH appropriately
If the server admin didn't specify a donation address, the software just uses
the author's adderss as a placeholder.

This commit adds code to transofrm this address to legacy BTC in case the
server is connected to a BTC bitcoind. If the user specified a donation
address, it is left unmolested (as before this commit).