Added batch request memory cost control and tracking
Just to prevent memory DoS or other nonsense, the batch requests cannot
exceed `max_buffer` in terms of the json data size for the both the
batch request itself + the results.

This tracking is done on a per-IP basis.  Batch requests that exceed the
`max_buffer` limit will be killed immediately.

As a batch request is processed, and as results accumulate, the memory cost
is updated to reflect the accumulated json data result byte size.

The memory costs are rough estimates, and are not exact, but they are good
enough approximations for the purposes of guarding against DoS.