Added --experimental-fast-sync CLI arg, also added more parallelism
We try ot prefetch for as long as possible before synchronizing the
addBlock thread to the prefetch worker.  As such we are forced to defer
writes to the UTXO cache until after the prefetch worker is done.

This might yield better parallelism, or it might not.

Also added CLI option to specify utxo cache size and enable this
newfangled utxo cache facility (off by default).

Also cleaned up the prefetcher code to persist its vectors/hashmaps
across calls to avoid redundant reallocations.