Batching: Allow `null` and duplicate `id`'s in batches
We removed the limitation we had previously on the `id`'s seen inside
bath request.  This was done by more correctly modeling the problem and
upgrading the internals of RPC::ConnectionBase to correctly associate
each RPC message's request <-> response with its execution context.

This did involve passing down the BatchId's everywhere, but that's the
correct solution anyway, and ended up simplifying a lot of internals and
removing some corner case strangeness.