Made the 'RecordFile' class always encode its header in little endian
This class now encodes the flat file header in little endian always (on
little endian systems this was always the case).  Now, should Fulcrum happen
to be compiled on a big endian system, this class will always encode in
little endian to the header, which makes the "RecordFile" flat-file db's
platform-neutral.

This change may theoretically break compatibility with previous Fulcrum
installs that were compiled on a big endian system, since before this
change the RecordFiles always wrote to the header using platform-specific
endianness. Code has been added to "upgrade" big endian db's to use
little endian, however this will not be a backward compatible upgrade.

However, this is highly unlikely to break any real installs since probably
100% of Fulcrum installs are on little endian currently.

This commit is just for correctness going forward to move the db format
into a less platform-dependent format.