On startup, proceed anyway if PeerMgr fails to start
In issue #109, the user tried to start up a regtest Fulcrum instance and
was surprised when it immediately failed upon synching. This was
happening because the PeerMgr could not be started for regtest (since
such a thing makes no sense, really).  This was causing Fulcrum to fail
out and refuse to run.  This wouldn't normally be a problem if
`peering=false` is specified in the conf file, but the default is
`peering=true` so users may be surprised by this behavior of the app
just refusing to work on regtest, out of the box, without extra conf
file options being needed.

Rather than abort the app if the PeerMgr fails, we simply log it as an
error and continue.