Docker build: Add new build platform "linux_ub22" (Ubuntu Jammy)
We added support for building using an Ubuntu Jammy docker container. In
the future we may switch to this one as the primary, but for now it will
just be there as a testing option.

Also, this change may be useful in case we want to switch to using Qt6
(which supports OpenSSL 3, which is what Jammy ships with).