Change to the way --fast-sync is checked and enforced.
It turns out some users leave this option alive in their conf files, so
it can cause intermittent startup failures if the option happens to be
in the conf file, even after initial synch, and the machine happens to
lack enough available RAM at startup.

But after initial synch this option is ignored anyway so enforcing it at
startup is almost nonsensical.

The check has been moved later into the pipeline -- into Storage.cpp,
and it becomes a soft check, whereby if the --fast-sync option is there
it will just be capped at available physical RAM should it happen to
exceed it.

Fixes issue #208.