RPC: Added blockchain.*.get_first_use
This is currently implemented identically to the original Electrs
implementation of the method.

Note that we do not support the new Rostrum optional 2nd arg "token_filter"
for this method since we don't store historical token data in the db.

Also note that unlike Electrs, but like Rostrum, we do not add anything
to the features map to indicate support for this method. Instead,
protocol 1.5.2 or greater will support this method.

Also in this commit: bumped protocol version to 1.5.2.