Mempool bench: Added optional mechanism to download real coins from bitcoind
For `--bench mempool`, added the ability to specify a bitcoind daemon to
connect to (via env var BITCOIND), in addition to the mempool.dat file (env
var MPDAT). If BITCOIND is specified in env, then we don't synthesize
the confirmed coins being spent, instead we download the real prevTx coins
that are referenced by the mempool.dat.

It is hoped this will lead to a more realistic simulation when benching &
testing the mempool code of Fulcrum.