Windows: Bumped docker image to Ubuntu 20; mxe GCC to 12.2
For some reason, using gcc 11.2 with rocksdb 9.1.1 for the Windows
build led to a segfault in rocksdb. But bumping to use GCC 12.2
on Windows is ok.

So we updated the mxe commit hash to something newer whereby GCC 12.2
is included. This required a newer host system (Ubuntu 20), however.

Now, the windows build is ok again and works with rocksdb 9.1.1.