Rpa: Fast-path optimization for testnet(s)
Testnets have, and testnet4 in particular has, many empty blocks. So, we
implement a fast-path optimization for Rpa table serialization where the
empty prefix table, which always serializes to a bunch of zeroes
essentially, gets cached and re-used for empty prefix tables at runtime.

This saves cycles on testnet4 and chipnet synching.