Rpa: Code nit to not use QByteArray for Rpa::Hash but instead use prevector
This is a small optimization to use `prevector` for the `Rpa::Hash`
type, which avoids allocations in the common case. `QByteArray` always
allocates a buffer and in the common codepaths we don't really need to
do that, since the calculated hash value is small and ephemeral anyway,
and can very well live on the stack.

This should save cycles and help parallelism on block processing ever so
slightly.