Parallelize the RPA-related processing when processing blocks
We move all the RPA-related processing when processing a block in
`PreProcessedBlock::fill` to a `CoTask` to be done in another thread in
parallel. For initial synch this doesn't make too much of a difference
since initial synch is processing multiple blocks in parallel anyway,
and all cores should be more-or-less busy.

However, the payoff is anticipated more for one-off block downloads as new
blocks arrive. Previous to this commit, only 1 thread would be
filling-in the block data in `PreProcessedBlock::fill`. After this commit,
the RPA data is filled-in in parallel in a separate thread which should
hopefully make processing faster for new blocks arriving as the server is up
(after initial synch).