Minor code quality and perf. tweak to UTXO Cache Prefetcher
It was unnecessarily using a hash table to store TXOs, when a vector
would do, since we are processing them in order and the lookup was by
index anyway! Not sure why I originally used a hash table here but a
vector is definitely more appropriate here.  Should make prefetching
ever-so-slightly faster if using --utxo-cache on initial sync.