DBRecordArray: allow DB-specified and caller-specified bucketNRecs to disagree
In the case where there is a pre-existing db table and the metadata
claims a different bucketNRecs than what the caller specified, it's
better to just go with the DB value to maintain forward and backward
compatibility rather than error-out. So the DB-specified value is used,
and everything is peachy.