Protocol 1.6: Added mempool.get_info, and also a scheme to cache bitcoind responses
Added a new method, mempoool.get_info which is just the result of
`getmempoolinfo` on the bitcoind side.

We cache this response for 5 seconds. Added a generic mechanism to cache
bitcoind responses for calls that do not need to be super up-to-date and
which can be cached for a time.

The cache is owned by BitcoinDMgr and is cleared when re-connecting to
bticoind. It auto-cleans itself when entries get stale.