PeerMgr: On connection failure to TCP port, retry SSL port (if any)
This helps us tolerate server misconfigurations where the TCP port is
invalid/not-working but the SSL port would work. e.keff.org on BTC
mainnet is one such server.

We prefer to be tolerant here to pick up as many "good" servers as
possible, in the interests of wallets obtaining as many options as
possible for servers.

This is ok to do and preferable since wallets almost exclusively use the
SSL port anyway.